+++ sdk/resourcemanager/marketplace/armmarketplace/zz_generated_privatestorecollection_client.go
@@ -51,9 +51,62 @@ func NewPrivateStoreCollectionClient(credential azcore.TokenCredential, options 
 	return client, nil
 }
 
+// ApproveAllItems - Delete all existing offers from the collection and enable approve all items.
+// If the operation fails it returns an *azcore.ResponseError type.
+// Generated from API version 2022-03-01
+// privateStoreID - The store ID - must use the tenant ID
+// collectionID - The collection ID
+// options - PrivateStoreCollectionClientApproveAllItemsOptions contains the optional parameters for the PrivateStoreCollectionClient.ApproveAllItems
+// method.
+func (client *PrivateStoreCollectionClient) ApproveAllItems(ctx context.Context, privateStoreID string, collectionID string, options *PrivateStoreCollectionClientApproveAllItemsOptions) (PrivateStoreCollectionClientApproveAllItemsResponse, error) {
+	req, err := client.approveAllItemsCreateRequest(ctx, privateStoreID, collectionID, options)
+	if err != nil {
+		return PrivateStoreCollectionClientApproveAllItemsResponse{}, err
+	}
+	resp, err := client.pl.Do(req)
+	if err != nil {
+		return PrivateStoreCollectionClientApproveAllItemsResponse{}, err
+	}
+	if !runtime.HasStatusCode(resp, http.StatusOK) {
+		return PrivateStoreCollectionClientApproveAllItemsResponse{}, runtime.NewResponseError(resp)
+	}
+	return client.approveAllItemsHandleResponse(resp)
+}
+
+// approveAllItemsCreateRequest creates the ApproveAllItems request.
+func (client *PrivateStoreCollectionClient) approveAllItemsCreateRequest(ctx context.Context, privateStoreID string, collectionID string, options *PrivateStoreCollectionClientApproveAllItemsOptions) (*policy.Request, error) {
+	urlPath := "/providers/Microsoft.Marketplace/privateStores/{privateStoreId}/collections/{collectionId}/approveAllItems"
+	if privateStoreID == "" {
+		return nil, errors.New("parameter privateStoreID cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{privateStoreId}", url.PathEscape(privateStoreID))
+	if collectionID == "" {
+		return nil, errors.New("parameter collectionID cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{collectionId}", url.PathEscape(collectionID))
+	req, err := runtime.NewRequest(ctx, http.MethodPost, runtime.JoinPaths(client.host, urlPath))
+	if err != nil {
+		return nil, err
+	}
+	reqQP := req.Raw().URL.Query()
+	reqQP.Set("api-version", "2022-03-01")
+	req.Raw().URL.RawQuery = reqQP.Encode()
+	req.Raw().Header["Accept"] = []string{"application/json"}
+	return req, nil
+}
+
+// approveAllItemsHandleResponse handles the ApproveAllItems response.
+func (client *PrivateStoreCollectionClient) approveAllItemsHandleResponse(resp *http.Response) (PrivateStoreCollectionClientApproveAllItemsResponse, error) {
+	result := PrivateStoreCollectionClientApproveAllItemsResponse{}
+	if err := runtime.UnmarshalAsJSON(resp, &result.Collection); err != nil {
+		return PrivateStoreCollectionClientApproveAllItemsResponse{}, err
+	}
+	return result, nil
+}
+
 // CreateOrUpdate - Create or update private store collection
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2021-12-01
+// Generated from API version 2022-03-01
 // privateStoreID - The store ID - must use the tenant ID
 // collectionID - The collection ID
 // options - PrivateStoreCollectionClientCreateOrUpdateOptions contains the optional parameters for the PrivateStoreCollectionClient.CreateOrUpdate
@@ -89,7 +142,7 @@ func (client *PrivateStoreCollectionClient) createOrUpdateCreateRequest(ctx cont
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-12-01")
+	reqQP.Set("api-version", "2022-03-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if options != nil && options.Payload != nil {
@@ -109,7 +162,7 @@ func (client *PrivateStoreCollectionClient) createOrUpdateHandleResponse(resp *h
 
 // Delete - Delete a collection from the given private store.
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2021-12-01
+// Generated from API version 2022-03-01
 // privateStoreID - The store ID - must use the tenant ID
 // collectionID - The collection ID
 // options - PrivateStoreCollectionClientDeleteOptions contains the optional parameters for the PrivateStoreCollectionClient.Delete
@@ -145,15 +198,68 @@ func (client *PrivateStoreCollectionClient) deleteCreateRequest(ctx context.Cont
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-12-01")
+	reqQP.Set("api-version", "2022-03-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
 }
 
+// DisableApproveAllItems - Disable approve all items for the collection.
+// If the operation fails it returns an *azcore.ResponseError type.
+// Generated from API version 2022-03-01
+// privateStoreID - The store ID - must use the tenant ID
+// collectionID - The collection ID
+// options - PrivateStoreCollectionClientDisableApproveAllItemsOptions contains the optional parameters for the PrivateStoreCollectionClient.DisableApproveAllItems
+// method.
+func (client *PrivateStoreCollectionClient) DisableApproveAllItems(ctx context.Context, privateStoreID string, collectionID string, options *PrivateStoreCollectionClientDisableApproveAllItemsOptions) (PrivateStoreCollectionClientDisableApproveAllItemsResponse, error) {
+	req, err := client.disableApproveAllItemsCreateRequest(ctx, privateStoreID, collectionID, options)
+	if err != nil {
+		return PrivateStoreCollectionClientDisableApproveAllItemsResponse{}, err
+	}
+	resp, err := client.pl.Do(req)
+	if err != nil {
+		return PrivateStoreCollectionClientDisableApproveAllItemsResponse{}, err
+	}
+	if !runtime.HasStatusCode(resp, http.StatusOK) {
+		return PrivateStoreCollectionClientDisableApproveAllItemsResponse{}, runtime.NewResponseError(resp)
+	}
+	return client.disableApproveAllItemsHandleResponse(resp)
+}
+
+// disableApproveAllItemsCreateRequest creates the DisableApproveAllItems request.
+func (client *PrivateStoreCollectionClient) disableApproveAllItemsCreateRequest(ctx context.Context, privateStoreID string, collectionID string, options *PrivateStoreCollectionClientDisableApproveAllItemsOptions) (*policy.Request, error) {
+	urlPath := "/providers/Microsoft.Marketplace/privateStores/{privateStoreId}/collections/{collectionId}/disableApproveAllItems"
+	if privateStoreID == "" {
+		return nil, errors.New("parameter privateStoreID cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{privateStoreId}", url.PathEscape(privateStoreID))
+	if collectionID == "" {
+		return nil, errors.New("parameter collectionID cannot be empty")
+	}
+	urlPath = strings.ReplaceAll(urlPath, "{collectionId}", url.PathEscape(collectionID))
+	req, err := runtime.NewRequest(ctx, http.MethodPost, runtime.JoinPaths(client.host, urlPath))
+	if err != nil {
+		return nil, err
+	}
+	reqQP := req.Raw().URL.Query()
+	reqQP.Set("api-version", "2022-03-01")
+	req.Raw().URL.RawQuery = reqQP.Encode()
+	req.Raw().Header["Accept"] = []string{"application/json"}
+	return req, nil
+}
+
+// disableApproveAllItemsHandleResponse handles the DisableApproveAllItems response.
+func (client *PrivateStoreCollectionClient) disableApproveAllItemsHandleResponse(resp *http.Response) (PrivateStoreCollectionClientDisableApproveAllItemsResponse, error) {
+	result := PrivateStoreCollectionClientDisableApproveAllItemsResponse{}
+	if err := runtime.UnmarshalAsJSON(resp, &result.Collection); err != nil {
+		return PrivateStoreCollectionClientDisableApproveAllItemsResponse{}, err
+	}
+	return result, nil
+}
+
 // Get - Gets private store collection
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2021-12-01
+// Generated from API version 2022-03-01
 // privateStoreID - The store ID - must use the tenant ID
 // collectionID - The collection ID
 // options - PrivateStoreCollectionClientGetOptions contains the optional parameters for the PrivateStoreCollectionClient.Get
@@ -189,7 +295,7 @@ func (client *PrivateStoreCollectionClient) getCreateRequest(ctx context.Context
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-12-01")
+	reqQP.Set("api-version", "2022-03-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -206,7 +312,7 @@ func (client *PrivateStoreCollectionClient) getHandleResponse(resp *http.Respons
 
 // List - Gets private store collections list
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2021-12-01
+// Generated from API version 2022-03-01
 // privateStoreID - The store ID - must use the tenant ID
 // options - PrivateStoreCollectionClientListOptions contains the optional parameters for the PrivateStoreCollectionClient.List
 // method.
@@ -237,7 +343,7 @@ func (client *PrivateStoreCollectionClient) listCreateRequest(ctx context.Contex
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-12-01")
+	reqQP.Set("api-version", "2022-03-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -254,7 +360,7 @@ func (client *PrivateStoreCollectionClient) listHandleResponse(resp *http.Respon
 
 // Post - Delete Private store collection. This is a workaround.
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2021-12-01
+// Generated from API version 2022-03-01
 // privateStoreID - The store ID - must use the tenant ID
 // collectionID - The collection ID
 // options - PrivateStoreCollectionClientPostOptions contains the optional parameters for the PrivateStoreCollectionClient.Post
@@ -290,7 +396,7 @@ func (client *PrivateStoreCollectionClient) postCreateRequest(ctx context.Contex
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-12-01")
+	reqQP.Set("api-version", "2022-03-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if options != nil && options.Payload != nil {
@@ -301,7 +407,7 @@ func (client *PrivateStoreCollectionClient) postCreateRequest(ctx context.Contex
 
 // TransferOffers - transferring offers (copy or move) from source collection to target collection(s)
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2021-12-01
+// Generated from API version 2022-03-01
 // privateStoreID - The store ID - must use the tenant ID
 // collectionID - The collection ID
 // options - PrivateStoreCollectionClientTransferOffersOptions contains the optional parameters for the PrivateStoreCollectionClient.TransferOffers
@@ -337,7 +443,7 @@ func (client *PrivateStoreCollectionClient) transferOffersCreateRequest(ctx cont
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-12-01")
+	reqQP.Set("api-version", "2022-03-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	if options != nil && options.Payload != nil {